refactor: 결제 금액 데이터 타입 BigDecimal로 변경

+++ src/main/java/roomescape/dto/reservation/UserReservationPaymentRequest.java
@@ -1,6 +1,6 @@
 package roomescape.dto.reservation;
 
-import com.fasterxml.jackson.annotation.JsonFormat;
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record UserReservationPaymentRequest(
@@ -10,7 +10,7 @@ public record UserReservationPaymentRequest(
         Long memberId,
         String paymentKey,
         String orderId,
-        int amount,
+        BigDecimal amount,
         String paymentType
 ) {
 }
